refactor: refactoring usercontroller for user name input
