+++ docs/readme.md
@@ -5,10 +5,10 @@
 - [x] 공백으로 이루어진 이름
 - [x] 이름의 시작이 공백
 - [x] 중복되는 이름
+- [x] 5자 이하의 이름
+      [ ]사용자는 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
 
-[ ]사용자는 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
-
-- [ ] 정수의 횟수
+- [ ]정수의 횟수
 - [ ] 0이상의 횟수
 
 [ ]주어진 횟수 동안 n 대의 자동차는 전진 또는 멈출 수 있다.
