+++ src/index.js
@@ -1,8 +1,8 @@
-import { setCarNames } from './controllers/userControllers.js';
+import { setCarNamesClick } from './controllers/userControllers.js';
 
 const state = {
   cars: '',
   location: 0,
 };
 
-setCarNames(state);
+setCarNamesClick(state);
